+++ rqalpha/mod/simulation/simulation_broker.py
@@ -159,10 +159,12 @@ class SimulationBroker(AbstractBroker, Persistable):
         self._matcher.update(env.calendar_dt, env.trading_dt, bar_dict)
         self._match()
 
-    def tick(self):
-        env = Environment.get_instance()
-        self._matcher.update(env.calendar_dt, env.trading_dt, bar_dict)
-        self._match()
+    def tick(self, tick):
+        # TODO support tick matching
+        pass
+        # env = Environment.get_instance()
+        # self._matcher.update(env.calendar_dt, env.trading_dt, tick)
+        # self._match()
 
     def _match(self):
         self._matcher.match(self._open_orders)
@@ -170,7 +172,5 @@ class SimulationBroker(AbstractBroker, Persistable):
         self._open_orders = [(a, o) for a, o in self._open_orders if not o._is_final()]
 
         for account, order in final_orders:
-            if order.status == ORDER_STATUS.REJECTED:
+            if order.status == ORDER_STATUS.REJECTED or order.status == ORDER_STATUS.CANCELLED:
                 self._env.event_bus.publish_event(EVENT.ORDER_UNSOLICITED_UPDATE, account, order)
-            elif order.status == ORDER_STATUS.CANCELLED:
-                self._env.event_bus.publish_event(EVENT.ORDER_CANCELLATION_PASS, account, order)