+++ rqalpha/mod/simulation/simulation_event_source.py
@@ -83,6 +83,7 @@ class SimulationEventSource(AbstractEventSource):
 
     def events(self, start_date, end_date, frequency):
         if frequency == "1d":
+            # 根据起始日期和结束日期，获取所有的交易日，然后再循环获取每一个交易日
             for day in self._env.data_proxy.get_trading_dates(start_date, end_date):
                 date = day.to_pydatetime()
                 dt_before_trading = date.replace(hour=0, minute=0)
@@ -120,7 +121,8 @@ class SimulationEventSource(AbstractEventSource):
                             trading_dt = calendar_dt
                         if before_trading_flag:
                             before_trading_flag = False
-                            yield Event(EVENT.BEFORE_TRADING, trading_dt, trading_dt)
+                            before_trading_dt = trading_dt - datetime.timedelta(minutes=30)
+                            yield Event(EVENT.BEFORE_TRADING, before_trading_dt, before_trading_dt)
                         if self._universe_changed:
                             self._universe_changed = False
                             last_dt = calendar_dt