+++ rqalpha/model/account/future_account.py
@@ -99,8 +99,8 @@ class FutureAccount(BaseAccount):
         portfolio = self.portfolio
         portfolio._portfolio_value = None
         position = portfolio.positions[tick.order_book_id]
-        position._last_price = tick.last_price
-        self._update_market_value(position, tick.last_price)
+        position._last_price = tick.last
+        self._update_market_value(position, tick.last)
 
     def order_pending_new(self, account, order):
         if self != account:
@@ -124,6 +124,7 @@ class FutureAccount(BaseAccount):
             return
         order_book_id = order.order_book_id
         position = self.portfolio.positions[order_book_id]
+        position._total_orders -= 1
         cancel_quantity = order.unfilled_quantity
         cancel_value = -order._frozen_price * cancel_quantity * position._contract_multiplier
         frozen_margin = self.margin_decider.cal_margin(order_book_id, order.side, cancel_value)
@@ -134,18 +135,24 @@ class FutureAccount(BaseAccount):
         pass
 
     def order_cancellation_pass(self, account, order):
-        if self != account:
-            return
-        order_book_id = order.order_book_id
-        position = self.portfolio.positions[order_book_id]
-        canceled_quantity = order.unfilled_quantity
-        canceled_value = -order._frozen_price * canceled_quantity * position._contract_multiplier
-        frozen_margin = self.margin_decider.cal_margin(order_book_id, order.order_book_id, canceled_value)
-        self._update_order_data(position, order, -canceled_quantity, canceled_value)
-        self._update_frozen_cash(order, frozen_margin)
+        self._cancel_order_cal(account, order)
 
     def order_cancellation_reject(self, account, order):
         pass
+
+    def order_unsolicited_update(self, account, order):
+        self._cancel_order_cal(account, order)
+
+    def _cancel_order_cal(self, account, order):
+        if self != account:
+            return
+        order_book_id = order.order_book_id
+        position = self.portfolio.positions[order.order_book_id]
+        rejected_quantity = order.unfilled_quantity
+        rejected_value = -order._frozen_price * rejected_quantity * position._contract_multiplier
+        frozen_margin = self.margin_decider.cal_margin(order_book_id, order.side, rejected_value)
+        self._update_order_data(position, order, -rejected_quantity, rejected_value)
+        self._update_frozen_cash(order, frozen_margin)
 
     def trade(self, account, trade):
         if self != account:
@@ -182,17 +189,7 @@ class FutureAccount(BaseAccount):
         self._update_trade_data(position, trade, trade_quantity, trade_value)
 
         self._update_market_value(position, bar_dict[order_book_id].close)
-
-    def order_unsolicited_update(self, account, order):
-        if self != account:
-            return
-        order_book_id = order.order_book_id
-        position = self.portfolio.positions[order.order_book_id]
-        rejected_quantity = order.unfilled_quantity
-        rejected_value = -order._frozen_price * rejected_quantity * position._contract_multiplier
-        frozen_margin = self.margin_decider.cal_margin(order_book_id, order.side, rejected_value)
-        self._update_order_data(position, order, -rejected_quantity, rejected_value)
-        self._update_frozen_cash(order, frozen_margin)
+        self._last_trade_id = trade.exec_id
 
     @staticmethod
     def _update_holding_by_settle(position, settle_price):